Move Cache from ember-metal to ember-utils.
Also remove usage of `UNDEFINED` symbol (just use Map behaviors).
