+++ packages/@ember/string/index.js
@@ -5,7 +5,7 @@
 export { getStrings as _getStrings, setStrings as _setStrings } from './lib/string_registry';
 
 import { ENV } from 'ember-environment';
-import { Cache } from 'ember-metal';
+import { Cache } from 'ember-utils';
 import { getString } from './lib/string_registry';
 
 const STRING_DASHERIZE_REGEXP = /[ _]/g;
