+++ packages/ember-metal/index.js
@@ -9,7 +9,6 @@ export {
 export { default as alias } from './lib/alias';
 export { deprecateProperty } from './lib/deprecate_property';
 export { descriptorFor, meta, peekMeta, deleteMeta } from './lib/meta';
-export { default as Cache } from './lib/cache';
 export { PROXY_CONTENT, _getPath, get, getWithDefault } from './lib/property_get';
 export { set, trySet } from './lib/property_set';
 export {
